Install example "etc" files into a package's "etc" directory.
Make postinstall scripts copy from the package's "etc" directory
to the ${ENCAP_TARGET}/etc directory.  This makes is more clear
where files are supposed to go to someone inspecting the package
directory.  Exclude the "etc" directory from linking by adding it
to the encapinfo files where needed.

Also fix the path to perl in the openssl tsget script.

Bump the pkgver of openldap-client, openslp and openssl.