Add support for "pecan_tools_build" and "pecan_tools_pkg".
These two variables take lists of tools and automatically add
prereq for these tools if needed.  This improves the tools
support in Pecan by allowing a package pecan.sh file to state
which tools it needs during build or in the final package without
explicitly listing the prereq.  If the prereq added automatically
is insufficient, then the pecan.sh script can fall back to
explicitly using pecan_prereq_{build,pkg} for that tool.

Bump the pecan_build_version to 2 for the new features.