Leave ``pecan_srcdir'' as top of the extracted tree.
Create new variables to represent the directory in which we configure,
build, test and install:

	pecan_{configure,build,test,install}_srcdir

``pecan_configure_srcdir'' defaults to ${pecan_srcdir} and each
of the other variables defaults to the value set in its action
prerequisite.

Patch files are now always relative to ${pecan_srcdir}.  This works
around an issue where GNU patch won't patch a file specified with
"../" in its path.