pecan: Support ABI versions of library packages.
Every package that installs shared libraries stores a new
``abi_version'' value in its encapinfo.  This ABI version is the
lowest version of the package which has the same ABI.

Modify the recorded library prereqs in package encapinfo files to
require at least the ABI version of those packages.

pecan.sh files can now define two new variables ``pecan_abi_version''
and ``pecan_api_version'' that hold the ABI version of the package,
and the suggested version of the package to use when constructing a
library prereq for a package.

Bump ``pecan_build_version'' to 8.